+++ src/features/create-new-password/model/createNewPasswordFormValues.ts
@@ -1,0 +1,4 @@
+export type CreateNewPasswordFormValues = {
+  newPassword: string
+  passwordConfirmation: string
+}